--- a/src/core/Akka.Remote/Configuration/Remote.conf
+++ b/src/core/Akka.Remote/Configuration/Remote.conf
@@ -432,11 +432,6 @@ akka {
       # i.e. how long a connect may take until it is timed out
       connection-timeout = 15 s
 
-      # Toggles buffer pooling on and off inside DotNetty.
-      # Only intended to be a work-around for users who are still running on DotNetty v0.4.6-v0.4.7
-      # for the following bug: https://github.com/akkadotnet/akka.net/issues/3370
-      enable-pooling = true
-
       # PERFORMANCE TUNING
       # 
       # The batching feature of DotNetty is designed to help batch together logical writes into a smaller