@dnickless 2020-02-26
I actually do see a major difference in the numbers... The first line for a single client appears to be off by some serious factor?