@Aaronontheweb 2020-02-26
@dnickless that's a byproduct of our batching system that we added, which improved overall performance significantly: https://github.com/akkadotnet/akka.net/pull/4106

The first round of the benchmark puts very little pressure on either side of the wire since it's a single actor emitting 20 messages at a time, beneath the default threshold of 30 messages per batch used by the batching system. As a result, we're depending on the recurring 40ms timer to flush out writes in that scenario - hence why the first round of the benchmark is a big outlier, since the operating system can't guarantee _exactly_ 40ms each time. I added an article on how to performance tune this new batching system to the Akka.NET documentation here: https://getakka.net/articles/remoting/performance.html

Just to give you some idea on the numbers:

Before: https://getakka.net/articles/remoting/performance.html#no-io-batching

Average performance: 82,539 msg/s.

Standard deviation: 46,827 msg/s.

After: https://getakka.net/articles/remoting/performance.html#with-io-batching

Average performance: 141,091 msg/s.

Standard deviation: 15,291 msg/s.

Those figures came from a different piece of hardware than the one I used for the benchmark yesterday (we haven't replaced our standard benchmarking setup since moving onto Azure DevOps instead of our home-grown CI.) The figures on the website came from a 2019-generation development laptop. The figures on this PR came from a 2012 laptop running much older hardware.

Just now, I re-ran the benchmark with these changes on a third machine (home office machine - AMD Ryzen setup built in 2017) which has overall better hardware.

```
λ  dotnet run -c Release --framework netcoreapp3.1
ProcessorCount:                    16
ClockSpeed:                        0 MHZ
Actor Count:                       32
Messages sent/received per client: 200000  (2e5)
Is Server GC:                      True

Num clients, Total [msg], Msgs/sec, Total [ms]
         1,  200000,     58755,    3404.19
         5, 1000000,    183790,    5441.41
        10, 2000000,    178779,   11187.50
        15, 3000000,    180452,   16625.69
        20, 4000000,    178277,   22437.16
        25, 5000000,    180291,   27733.56
```

You'll see the big drop in the first round of performance there too - it's the same root issue: falling beneath the batching threshold. All of those batch thresholds (time, msg count, total bytes) can be customized according to the article I linked earlier in Akka.NET v1.4.0.